+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog Fab-manager
 
+- OpenAPI spaces endpoints (index/show)
 - Improved automated test on statistics generation
 - Refactored statistics generation
 - Refactored test helpers